==> scripts/2_labeling.ipynb <==
import pandas as pd
from langdetect import detect, DetectorFactory, LangDetectException
from tqdm import tqdm
import os
DetectorFactory.seed = 42

df = pd.read_csv("C:/Users/rahel/OneDrive/Dokumen/asah_AnalisisSentimen/data/gojek_reviews_full.csv")
print(f"Data loaded: {len(df)} rows")
df.head()

def language_detect(text):
    try:
        return detect(text)
    except LangDetectException:
        return "unknown"

tqdm.pandas(desc="Detecting language")
df["language"] = df["content"].progress_apply(language_detect)

df = df[df["language"] == "id"]
print("Filtered to Indonesian reviews:", len(df), "rows remaining.")

def labeling(score):
    score = int(score)
    if score <=2:
        return "negative"
    elif score == 3:
        return "neutral"
    else:
        return "positive"

df["label"] = df["score"].progress_apply(labeling)
df["label"].value_counts()

os.makedirs("../data/processed", exist_ok=True)
output_path = "../data/processed/gojek_reviews_labeled.csv"
df.to_csv(output_path, index=False, encoding="utf-8-sig")

print(f"Labeled data saved to {output_path}, total rows: {len(df)}")
