==> scripts/3_preprocessing.ipynb <==
!pip install Sastrawi

!pip install nltk

import pandas as pd
import re
import string
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory
from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
import nltk

nltk.download('punkt')
nltk.download('stopwords')

df = pd.read_csv("C:\\Users\\rahel\\OneDrive\\Dokumen\\asah_AnalisisSentimen\\data\\processed\\gojek_reviews_labeled.csv")
df = df.dropna(subset=['content'])
df = df.reset_index(drop=True)

kamus = pd.read_csv("C:\\Users\\rahel\\OneDrive\\Dokumen\\asah_AnalisisSentimen\\scripts\\kamusalay.csv")
alay_dict = dict(zip(kamus['slang'], kamus['formal']))

factory = StemmerFactory()
stemmer = factory.create_stemmer()
stop_words = set(stopwords.words('indonesian'))

def clean_text(text):
    text = str(text).lower()
    text = re.sub(r'http\S+|www\S+|https\S+', '', text)
    text = re.sub(r'@\w+|\#','', text)
    text = re.sub(r'[^a-zA-Z\s]', '', text)
    text = re.sub(r'\s+', ' ', text).strip()
    return text

def normalize(text):
    words = text.split()
    normalized_words = [alay_dict.get(word, word) for word in words]
    return ' '.join(normalized_words)

def tokenize(text):
    return word_tokenize(text)

def stopwords_removal(tokens):
    return [word for word in tokens if word not in stop_words]

def stem(text):
    return [stemmer.stem(word) for word in text]

def preprocess_text(text):
    text = clean_text(text)
    text = normalize(text)
    tokens = tokenize(text)
    tokens = stopwords_removal(tokens)
    tokens = stem(tokens)
    return ' '.join(tokens)

import os

output_dir = "../data/processed"
os.makedirs(output_dir, exist_ok=True)

df["cleaned"] = df["content"].apply(preprocess_text)

output_path = os.path.join(output_dir, "gojek_reviews_preprocessed.csv")
df.to_csv(output_path, index=False)

print(f"Preprocessing completed and saved to: {output_path}")
