==> scripts/5_featurextraction.ipynb <==
!pip install scikit-learn

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import pandas as pd
import os

processed = '../data/processed/gojek_reviews_preprocessed.csv'
df = pd.read_csv(processed)
df.head()

df = df.dropna(subset=['cleaned']).reset_index(drop=True)


bow = CountVectorizer(max_features=1000)
bow_features = bow.fit_transform(df['cleaned'])
bow_df = pd.DataFrame(bow_features.toarray(), columns=bow.get_feature_names_out())

tfidf = TfidfVectorizer(max_features=1000)
tfidf_features = tfidf.fit_transform(df['cleaned'])
tfidf_df = pd.DataFrame(tfidf_features.toarray(), columns=tfidf.get_feature_names_out())


output_dir = '../data/processed'
os.makedirs(output_dir, exist_ok=True)

bow_df.to_csv(os.path.join(output_dir, 'bow_features.csv'), index=False)
tfidf_df.to_csv(os.path.join(output_dir, 'tfidf_features.csv'), index=False)
print("Feature extraction completed and files saved.")
print("Bow features shape:", bow_df.shape)
print("TF-IDF features shape:", tfidf_df.shape)
